Refactor: Modify the logic for user withdrawal

+++ src/api/privacy.ts
@@ -21,24 +21,27 @@ export const updateUserInfo = async (
   }
 };
 
-export const deleteUserAccount = async (): Promise<void> => {
+export const deleteUserAccount = async (userId: number): Promise<void> => {
   try {
     const token = localStorage.getItem("authToken");
+    const isAdmin = localStorage.getItem("is_admin");
     if (!token) {
       throw new Error("User not authenticated");
     }
 
-    await API.delete("/users", {
+    await API.delete(`/users/${userId}`, {
       headers: {
         Authorization: `Bearer ${token}`,
       },
     });
 
     // 스토리지에서 사용자 정보 삭제
-    localStorage.removeItem("authToken");
-    localStorage.removeItem("nickname");
-    localStorage.removeItem("id");
-    localStorage.removeItem("email");
+    if (isAdmin === "false") {
+      localStorage.removeItem("authToken");
+      localStorage.removeItem("nickname");
+      localStorage.removeItem("id");
+      localStorage.removeItem("email");
+    }
   } catch (error: any) {
     throw new Error(
       error.response.data.message ||
